Refactor discussion n vote
