+++ web/pages/dashboard/governance.tsx
@@ -1,0 +1,10 @@
+import React from "react"
+import PageLayout from "../../components/layouts/PageLayout"
+
+export default function Governance() {
+  return (
+    <PageLayout containerClassName="bg-custom-blue bg-cover min-h-screen">
+      governance
+    </PageLayout>
+  )
+}
