+++ web/pages/dashboard/manage-sbt.tsx
@@ -1,0 +1,10 @@
+import React from "react"
+import PageLayout from "../../components/layouts/PageLayout"
+
+export default function ManageSbt() {
+  return (
+    <PageLayout containerClassName="bg-custom-blue bg-cover min-h-screen">
+      manage-sbt
+    </PageLayout>
+  )
+}
